+++ packages/dart_mappable_builder/lib/src/generators/record_mapper_generator.dart
@@ -42,6 +42,7 @@ class RecordMapperGenerator extends MapperGenerator<RecordMapperElement> {
 
     generateTypeFactory(output);
     generateApplyOverride(output);
+    generateHook(output);
     generateInstantiate(output);
     generateStaticDecoders(output);
 
@@ -129,5 +130,15 @@ class RecordMapperGenerator extends MapperGenerator<RecordMapperElement> {
       '    return ensureInitialized().decodeJson<${element.className}${element.typeParams}>(json);\n'
       '  }\n',
     );
+  }
+
+  void generateHook(StringBuffer output) {
+    var hook = element.hookForElement;
+    if (hook != null) {
+      output.write('''
+        @override
+        final MappingHook hook = $hook;
+      ''');
+    }
   }
 }